backend/requirements-test.txt
# Testing Requirements for TripCraft AI Backend
# Comprehensive testing framework dependencies

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
pytest-html>=3.2.0
pytest-json-report>=1.5.0

# HTTP Testing
httpx>=0.24.0
requests>=2.31.0
aiohttp>=3.8.0

# Test Data & Mocking
factory-boy>=3.3.0
faker>=19.0.0
responses>=0.23.0
pytest-mock>=3.11.0

# Rich Console & Reporting
rich>=13.5.0
click>=8.1.0
colorama>=0.4.6

# Performance Testing
pytest-benchmark>=4.0.0
locust>=2.16.0

# Security Testing
bandit>=1.7.5
safety>=2.3.0

# Code Quality
flake8>=6.0.0
black>=23.7.0
isort>=5.12.0
mypy>=1.5.0

# Coverage & Reporting
coverage>=7.3.0
pytest-cov>=4.1.0

# File Processing (for test data)
PyPDF2>=3.0.0
Pillow>=10.0.0

# Database Testing
sqlalchemy>=2.0.0
alembic>=1.12.0

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.0.0

# Async Testing Utilities
asyncio-mqtt>=0.13.0
aiofiles>=23.2.0

# Chart Generation (for reports)
matplotlib>=3.7.0
plotly>=5.15.0

# Documentation Testing
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Load Testing
artillery>=2.0.0  # Note: This is a Node.js package, included for reference

# API Documentation Testing
openapi-spec-validator>=0.6.0
jsonschema>=4.19.0

# Time & Date Testing
freezegun>=1.2.0
python-dateutil>=2.8.0

# Logging & Monitoring
structlog>=23.1.0
prometheus-client>=0.17.0

# Network Testing
pytest-httpserver>=1.0.8
wiremock>=2.6.0

# Parallel Testing
pytest-parallel>=0.1.1
pytest-xdist>=3.3.0

# Test Fixtures & Utilities
pytest-fixtures>=0.1.0
pytest-factoryboy>=2.5.0

# Memory & Resource Testing
psutil>=5.9.0
memory-profiler>=0.61.0

# Container Testing (for Docker tests)
testcontainers>=3.7.0
docker>=6.1.0

# API Schema Testing
openapi3>=1.8.0
jsonref>=1.1.0

# Snapshot Testing
syrupy>=4.0.0

# Property-based Testing
hypothesis>=6.82.0

# Visual Testing (for HTML reports)
selenium>=4.11.0
playwright>=1.37.0

# Test Reporting & Analytics
allure-pytest>=2.13.0
pytest-html>=3.2.0

# Environment & Configuration Testing
pytest-env>=0.8.2
pytest-dotenv>=0.5.2

# Retry & Flaky Test Handling
pytest-rerunfailures>=12.0
flaky>=3.7.0

# Test Ordering & Dependencies
pytest-order>=1.1.0
pytest-dependency>=0.5.1

# Custom Assertions
assertpy>=1.1

# Test Data Generation
mimesis>=11.1.0
polyfactory>=2.8.0

# API Client Testing
respx>=0.20.0
httpretty>=1.1.4

# Time-based Testing
time-machine>=2.12.0
pytest-freezegun>=0.4.2

# Subprocess Testing
pytest-subprocess>=1.5.0

# Temporary File Testing
pytest-tmp-path>=0.1.0

# Plugin Development
pluggy>=1.3.0

# Test Discovery & Collection
pytest-collect-formatter>=0.3.0

# Test Metadata
pytest-metadata>=3.0.0

# Test Timeouts
pytest-timeout>=2.1.0

# Test Markers
pytest-marker-bugzilla>=0.9.0

# Test Reporting Extensions
pytest-json>=0.4.0
pytest-csv>=3.0.0

# Development Dependencies
ipython>=8.14.0
jupyter>=1.0.0

# Debugging
pdb++>=0.10.3
pytest-pdb>=0.2.0

# Test Configuration
toml>=0.10.2
pyyaml>=6.0.1
